:hammer: chore: refactor state & files
