+++ packages/client/locales/en/leo.json
@@ -193,6 +193,8 @@
     "award": "Award",
     "activeCall": "Active Call",
     "arrestReportLogs": "Arrest Report Logs",
+    "activeUnits": "Active Units",
+    "noActiveUnits": "There are no active units on the live map.",
     "alert_deleteQualification": "Are you sure you want to delete this qualification? This action cannot be undone.",
     "alert_deleteDLExam": "Are you sure you want to delete this driver's license exam? This action cannot be undone.",
     "vehicleImpoundLocation": "The provided vehicle ({plate}) will be impounded at {impoundLocation}. You will be able to view this vehicle on the impound lot page.",
